Add checkstyle

+++ extractor/src/main/java/org/schabi/newpipe/extractor/stream/Frameset.java
@@ -21,7 +21,7 @@ public final class Frameset implements Serializable {
             final int durationPerFrame,
             final int framesPerPageX,
             final int framesPerPageY) {
-        
+
         this.urls = urls;
         this.totalCount = totalCount;
         this.durationPerFrame = durationPerFrame;
@@ -84,8 +84,8 @@ public final class Frameset implements Serializable {
      * Returns the information for the frame at stream position.
      *
      * @param position Position in milliseconds
-     * @return An <code>int</code>-array containing the bounds and URL where the indexes are specified as
-     * followed:
+     * @return An <code>int</code>-array containing the bounds and URL where the indexes are
+     * specified as follows:
      *
      * <ul>
      *     <li><code>0</code>: Index of the URL</li>
@@ -96,9 +96,9 @@ public final class Frameset implements Serializable {
      * </ul>
      */
     public int[] getFrameBoundsAt(final long position) {
-        if (position < 0 || position > ((totalCount + 1) * durationPerFrame)) {
+        if (position < 0 || position > ((long) (totalCount + 1) * durationPerFrame)) {
             // Return the first frame as fallback
-            return new int[] { 0, 0, 0, frameWidth, frameHeight };
+            return new int[] {0, 0, 0, frameWidth, frameHeight};
         }
 
         final int framesPerStoryboard = framesPerPageX * framesPerPageY;
